requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-cors-headers==3.2.1
django-mysql==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.0
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface==0.6
Markdown==3.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.2.0.32
opt-einsum==3.1.0
packaging==20.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.6
python-dotenv==0.10.5
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
scipy==1.4.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
